Create Data for Stan Model

Create Data for Stan Model

Usage

make_stan_data(
  inf_model,
  obs_model,
  data,
  init_probs,
  epsilon = 1e-10,
  ih_cov = NULL,
  eh_cov = NULL
)

Arguments

inf_model

Infection process model object generated by make_infection_model

obs_model

Observation process model object generated by make_observation_model

data

data frame with columns the following columns:

"t", time of observation

"hh_id", numeric household ID which must be 1, 2, ..., N where N is the total number of household

"part_id", participant ID within the household which must be 1, 2, ... n where n is the number of observed household members

"hh_size", number of household members (whether observed or not)

At least one column with binary outcome information where the column name matches the outcome name provided to make_observation_model

init_probs

vector of initial probabilities for each infection process model state

epsilon

very small number to use for zero probability transitions

ih_cov

NULL for run without covariates, otherwise data frame with intra-household covariates for each participant

eh_cov

NULL for run without covariates, otherwise data frame with extra-household covariates for each participant

Value

A list with data elements for input to stan model.