Run Household Transmission Model

Run Household Transmission Model

Usage

run_model(
  inf_model,
  obs_model,
  data,
  init_probs,
  epsilon = 1e-10,
  ih_cov = NULL,
  eh_cov = NULL,
  iter = 2000,
  chains = 4,
  cores = getOption("mc.cores", 1L),
  init = NULL,
  save_states = FALSE
)

Arguments

inf_model

Infection process model object generated by make_infection_model

obs_model

Observation process model object generated by make_observation_model

data

data frame with columns the following columns:

"t", time of observation

"hh_id", numeric household ID which must be 1, 2, ..., N where N is the total number of household

"part_id", participant ID within the household which must be 1, 2, ... n where n is the number of observed household members

"hh_size", number of household members (whether observed or not)

At least one column with binary outcome information where the column name matches the outcome name provided to make_observation_model

init_probs

vector of initial probabilities for each infection process model state

epsilon

very small number to use for zero probability transitions

ih_cov

NULL for run without covariates, otherwise data frame with intra-household covariates for each participant

eh_cov

NULL for run without covariates, otherwise data frame with extra-household covariates for each participant

iter

number of MCMC iterations

chains

number of MCMC chains

cores

number of cores for parallelization

init

initial conditions for MCMC chains

save_states

indicator for whether to save state probabilities

file

file path to stan model

Value

draws_array object with chains for each model parameter

Examples

# Subset sir package data to 10 households (for speedy example)
sir_sub <- sir[sir$hh_id <= 10, ]

# Make infetion process model for SIR model
inf_mod <- make_infection_model(
  transmit(from = "S", to = "I"),
  progress(from = "I", to = "R", gamma = NA))

# Make observation process model
obs_mod <- make_observation_model(
  pcr = c("S" = 0.05, "I" = 0.95, "R" = 0.05),
  igg = c("S" = 0.01, "I" = 0.01, "R" = 0.8))

run_model(inf_model = inf_mod,
          obs_model = obs_mod,
          data = sir_sub,
          init_probs = c(1 - 2 * 1e-10, 1e-10, 1e-10))
#> 
#> SAMPLING FOR MODEL 'hmm' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.026733 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 267.33 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 113.672 seconds (Warm-up)
#> Chain 1:                105.234 seconds (Sampling)
#> Chain 1:                218.906 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'hmm' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.016621 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 166.21 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 115.26 seconds (Warm-up)
#> Chain 2:                102.085 seconds (Sampling)
#> Chain 2:                217.345 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'hmm' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.016741 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 167.41 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 108.959 seconds (Warm-up)
#> Chain 3:                112.558 seconds (Sampling)
#> Chain 3:                221.517 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'hmm' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.016625 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 166.25 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 119.269 seconds (Warm-up)
#> Chain 4:                117.81 seconds (Sampling)
#> Chain 4:                237.079 seconds (Total)
#> Chain 4: 
#> # A draws_array: 1000 iterations, 4 chains, and 3 variables
#> , , variable = eh_prob
#> 
#>          chain
#> iteration    1    2    3    4
#>         1 -3.9 -3.9 -4.4 -4.1
#>         2 -3.6 -4.0 -4.0 -4.1
#>         3 -4.0 -3.9 -3.7 -4.3
#>         4 -4.1 -3.4 -3.8 -3.9
#>         5 -4.2 -3.7 -4.1 -4.1
#> 
#> , , variable = ih_prob
#> 
#>          chain
#> iteration    1    2    3    4
#>         1 -2.8 -2.7 -2.6 -2.4
#>         2 -2.7 -3.0 -2.6 -2.5
#>         3 -2.9 -3.0 -3.4 -2.9
#>         4 -2.8 -3.4 -3.1 -2.5
#>         5 -3.1 -3.2 -2.8 -2.9
#> 
#> , , variable = gamma
#> 
#>          chain
#> iteration    1    2    3    4
#>         1 -1.8 -1.8 -1.8 -1.5
#>         2 -1.5 -1.7 -1.6 -1.7
#>         3 -1.4 -1.9 -1.5 -2.0
#>         4 -2.1 -1.9 -1.7 -1.4
#>         5 -1.7 -1.5 -1.4 -1.7
#> 
#> # ... with 995 more iterations